+++ packages/next-sitemap/src/index.ts
@@ -3,7 +3,7 @@ import { loadConfig } from './config'
 import { loadManifest } from './manifest'
 import { createUrlSet, generateUrl } from './url'
 import { buildSitemapXml } from './build-sitemap-xml'
-import { exportFile } from './export'
+import { exportFile } from './file'
 import { toChunks } from './array'
 import { resolveSitemapChunks, KNOWN_PATHS } from './path'
 import { generateRobotsTxt } from './robots-txt'
